Refactored directories inside lib and combined them into http
