--- lib/fs/constants.go
@@ -1,7 +1,0 @@
-package fs
-
-const (
-	CHUNK_SIZE = 1024
-	FOLDER_TYPE_PATH = "Folder"
-	FILE_TYPE_PATH = "File"
-)
