+++ lib/http/constants.go
@@ -8,11 +8,15 @@ const (
 	HEADER_LINE_SEPERATOR = "\r\n"
 	REQUEST_LINE_SEPERATOR = " "
 	HEADER_KEY_VALUE_SEPERATOR = ":"
-	MAX_VERSION = "HTTP/1.0"
+	MAX_VERSION = "1.0"
 	GET_METHOD = "GET"
 	POST_METHOD = "POST"
 	HEAD_METHOD = "HEAD"
 	DEFAULT_PORT = 8080
+	CHUNK_SIZE = 1024
+	FOLDER_TYPE_PATH = "Folder"
+	FILE_TYPE_PATH = "File"
+	VALIDATE_ROUTE_PATTERN = "^[a-zA-z][a-zA-Z0-9_/:-]*$"
 )
 
-var COMPATIBLE_VERSIONS = []string {"HTTP/1.0"}
+var COMPATIBLE_VERSIONS = []string {"1.0", "1.1"}
