+++ lib/http/filesystem.go
@@ -1,4 +1,4 @@
-package fs
+package http
 
 import (
 	"time"
