+++ lib/http/response.go
@@ -107,7 +107,7 @@ func (res *HttpResponse) writeStatusLine() error {
 		return errors.New("error occurred while writing response status line: Protocol version not set")
 	}
 
-	_, err := res.writer.WriteString(fmt.Sprintf("%s %d %s%s", res.Version, res.StatusCode, res.StatusMessage, HEADER_LINE_SEPERATOR))
+	_, err := res.writer.WriteString(fmt.Sprintf("HTTP/%s %d %s%s", res.Version, res.StatusCode, res.StatusMessage, HEADER_LINE_SEPERATOR))
 	if err != nil {
 		return errors.New("error occurred while writing response status line: " + err.Error())
 	}
