--- lib/router/constants.go
@@ -1,5 +1,0 @@
-package router
-
-const (
-	VALIDATE_ROUTE_PATTERN = "^[a-zA-z][a-zA-Z0-9_/:-]*$"
-)
