--- lib/router/utilities.go
@@ -1,24 +1,0 @@
-package router
-
-import (
-	"strings"
-	"regexp"
-)
-
-func validateRoute(Route string) bool {
-	if strings.HasPrefix(Route, "//") || !strings.HasPrefix(Route, "/") {
-		return false
-	}
-
-	RouteName := strings.TrimPrefix(Route, "/")
-	isRouteValid, err := regexp.MatchString(VALIDATE_ROUTE_PATTERN, RouteName)
-	if err != nil {
-		return false
-	}
-
-	if !isRouteValid {
-		return false
-	}
-	
-	return true
-}
